APPNG-2160 update lucene/tika

+++ appng-search/src/main/java/org/appng/search/indexer/DocumentIndexer.java
@@ -137,7 +137,7 @@ public class DocumentIndexer extends Consumer<DocumentEvent, DocumentProducer> i
 						BooleanQuery query = queryBuilder.build();
 						String queryString = query.toString();
 						TopDocs search = searcher.search(query, 10);
-						int found = search.totalHits;
+						long found = search.totalHits;
 						if (found > 0) {
 							indexWriter.deleteDocuments(query);
 							logger.debug("deleting " + found + " existing document(s) for query " + queryString);
